feat: added isIssuer
