+++ sdk/src/app/usecase/command/ssi/addIssuer/AddIssuerCommand.ts
