+++ sdk/src/app/usecase/command/ssi/error/AccountAlreadyInControlList.ts
@@ -203,47 +203,13 @@
 
 */
 
-import { QueryHandler } from '../../../../../core/decorator/QueryHandlerDecorator.js';
-import { IQueryHandler } from '../../../../../core/query/QueryHandler.js';
-import { RPCQueryAdapter } from '../../../../../port/out/rpc/RPCQueryAdapter.js';
-import { lazyInject } from '../../../../../core/decorator/LazyInjectDecorator.js';
-import SecurityService from '../../../../service/SecurityService.js';
-import { MirrorNodeAdapter } from '../../../../../port/out/mirror/MirrorNodeAdapter.js';
-import { HEDERA_FORMAT_ID_REGEX } from '../../../../../domain/context/shared/HederaId.js';
-import EvmAddress from '../../../../../domain/context/contract/EvmAddress.js';
-import {
-  GetIssuerListCountQuery,
-  GetIssuerListCountQueryResponse,
-} from './getIssuerListCountQuery.js';
+import BaseError, { ErrorCode } from '../../../../../core/error/BaseError.js';
 
-@QueryHandler(GetIssuerListCountQuery)
-export class GetIssuerListCountQueryHandler
-  implements IQueryHandler<GetIssuerListCountQuery>
-{
-  constructor(
-    @lazyInject(SecurityService)
-    public readonly securityService: SecurityService,
-    @lazyInject(MirrorNodeAdapter)
-    public readonly mirrorNodeAdapter: MirrorNodeAdapter,
-    @lazyInject(RPCQueryAdapter)
-    public readonly queryAdapter: RPCQueryAdapter,
-  ) {}
-
-  async execute(
-    query: GetIssuerListCountQuery,
-  ): Promise<GetIssuerListCountQueryResponse> {
-    const { securityId } = query;
-    const security = await this.securityService.get(securityId);
-    if (!security.evmDiamondAddress) throw new Error('Invalid security id');
-
-    const securityEvmAddress: EvmAddress = new EvmAddress(
-      HEDERA_FORMAT_ID_REGEX.exec(securityId)
-        ? (await this.mirrorNodeAdapter.getContractInfo(securityId)).evmAddress
-        : securityId.toString(),
+export class AccountIsAlreadyAnIssuer extends BaseError {
+  constructor(account: string) {
+    super(
+      ErrorCode.AccountIsAlreadyAnIssuer,
+      `The account ${account} is already an issuer`,
     );
-
-    const res = await this.queryAdapter.getIssuerListCount(securityEvmAddress);
-
-    return new GetIssuerListCountQueryResponse(res);
   }
 }
