+++ sdk/src/app/usecase/command/ssi/removeIssuer/RemoveIssuerCommand.ts
