+++ sdk/src/app/usecase/command/ssi/setRevocationRegistryAddress/SetRevocationRegistryAddressCommand.ts
