+++ sdk/src/app/usecase/query/ssi/getIssuerListCount/GetIssuerListCountQuery.ts
