+++ sdk/src/app/usecase/query/ssi/getRevocationRegistryAddress/GetRevocationRegistryAddressQuery.ts
