+++ sdk/src/app/usecase/query/ssi/isIssuer/IsIssuerQuery.ts
@@ -203,51 +203,18 @@
 
 */
 
-import {
-  GetIssuerListMembersQuery,
-  GetIssuerListMembersQueryResponse,
-} from './getIssuerListMembersQuery.js';
-import { QueryHandler } from '../../../../../core/decorator/QueryHandlerDecorator.js';
-import { IQueryHandler } from '../../../../../core/query/QueryHandler.js';
-import { RPCQueryAdapter } from '../../../../../port/out/rpc/RPCQueryAdapter.js';
-import { lazyInject } from '../../../../../core/decorator/LazyInjectDecorator.js';
-import SecurityService from '../../../../service/SecurityService.js';
-import { MirrorNodeAdapter } from '../../../../../port/out/mirror/MirrorNodeAdapter.js';
-import { HEDERA_FORMAT_ID_REGEX } from '../../../../../domain/context/shared/HederaId.js';
-import EvmAddress from '../../../../../domain/context/contract/EvmAddress.js';
+import { Query } from '../../../../../core/query/Query.js';
+import { QueryResponse } from '../../../../../core/query/QueryResponse.js';
 
-@QueryHandler(GetIssuerListMembersQuery)
-export class GetIssuerListMembersQueryHandler
-  implements IQueryHandler<GetIssuerListMembersQuery>
-{
+export class IsIssuerQueryResponse implements QueryResponse {
+  constructor(public readonly payload: boolean) {}
+}
+
+export class IsIssuerQuery extends Query<IsIssuerQueryResponse> {
   constructor(
-    @lazyInject(SecurityService)
-    public readonly securityService: SecurityService,
-    @lazyInject(MirrorNodeAdapter)
-    public readonly mirrorNodeAdapter: MirrorNodeAdapter,
-    @lazyInject(RPCQueryAdapter)
-    public readonly queryAdapter: RPCQueryAdapter,
-  ) {}
-
-  async execute(
-    query: GetIssuerListMembersQuery,
-  ): Promise<GetIssuerListMembersQueryResponse> {
-    const { securityId, start, end } = query;
-    const security = await this.securityService.get(securityId);
-    if (!security.evmDiamondAddress) throw new Error('Invalid security id');
-
-    const securityEvmAddress: EvmAddress = new EvmAddress(
-      HEDERA_FORMAT_ID_REGEX.exec(securityId)
-        ? (await this.mirrorNodeAdapter.getContractInfo(securityId)).evmAddress
-        : securityId.toString(),
-    );
-
-    const res = await this.queryAdapter.getIssuerListMembers(
-      securityEvmAddress,
-      start,
-      end,
-    );
-
-    return new GetIssuerListMembersQueryResponse(res);
+    public readonly securityId: string,
+    public readonly issuerId: string,
+  ) {
+    super();
   }
 }
