+++ sdk/src/app/usecase/query/ssi/isIssuer/IsIssuerQueryHandler.ts
@@ -203,32 +203,34 @@
 
 */
 
-import { CommandHandler } from '../../../../../core/decorator/CommandHandlerDecorator';
-import { AddIssuerCommand, AddIssuerCommandResponse } from './addIssuerCommand';
-import { ICommandHandler } from '../../../../../core/command/CommandHandler';
-import { lazyInject } from '../../../../../core/decorator/LazyInjectDecorator';
-import TransactionService from '../../../../service/TransactionService';
-import { MirrorNodeAdapter } from '../../../../../port/out/mirror/MirrorNodeAdapter';
-import EvmAddress from '../../../../../domain/context/contract/EvmAddress';
-import { HEDERA_FORMAT_ID_REGEX } from '../../../../../domain/context/shared/HederaId';
+import { QueryHandler } from '../../../../../core/decorator/QueryHandlerDecorator.js';
+import { IQueryHandler } from '../../../../../core/query/QueryHandler.js';
+import { RPCQueryAdapter } from '../../../../../port/out/rpc/RPCQueryAdapter.js';
+import { lazyInject } from '../../../../../core/decorator/LazyInjectDecorator.js';
+import SecurityService from '../../../../service/SecurityService.js';
+import { MirrorNodeAdapter } from '../../../../../port/out/mirror/MirrorNodeAdapter.js';
+import { HEDERA_FORMAT_ID_REGEX } from '../../../../../domain/context/shared/HederaId.js';
+import EvmAddress from '../../../../../domain/context/contract/EvmAddress.js';
+import { IsIssuerQuery, IsIssuerQueryResponse } from './IsIssuerQuery.js';
 
-@CommandHandler(AddIssuerCommand)
-export class AddIssuerCommandHandler
-  implements ICommandHandler<AddIssuerCommand>
-{
+@QueryHandler(IsIssuerQuery)
+export class IsIssuerQueryHandler implements IQueryHandler<IsIssuerQuery> {
   constructor(
-    @lazyInject(TransactionService)
-    public readonly transactionService: TransactionService,
+    @lazyInject(SecurityService)
+    public readonly securityService: SecurityService,
     @lazyInject(MirrorNodeAdapter)
-    private readonly mirrorNodeAdapter: MirrorNodeAdapter,
+    public readonly mirrorNodeAdapter: MirrorNodeAdapter,
+    @lazyInject(RPCQueryAdapter)
+    public readonly queryAdapter: RPCQueryAdapter,
   ) {}
 
-  async execute(command: AddIssuerCommand): Promise<AddIssuerCommandResponse> {
-    const { securityId, issuerId } = command;
-    const handler = this.transactionService.getHandler();
+  async execute(query: IsIssuerQuery): Promise<IsIssuerQueryResponse> {
+    const { securityId, issuerId } = query;
+    const security = await this.securityService.get(securityId);
+    if (!security.evmDiamondAddress) throw new Error('Invalid security id');
 
     const securityEvmAddress: EvmAddress = new EvmAddress(
-      HEDERA_FORMAT_ID_REGEX.test(securityId)
+      HEDERA_FORMAT_ID_REGEX.exec(securityId)
         ? (await this.mirrorNodeAdapter.getContractInfo(securityId)).evmAddress
         : securityId.toString(),
     );
@@ -237,14 +239,11 @@ export class AddIssuerCommandHandler
       ? await this.mirrorNodeAdapter.accountToEvmAddress(issuerId)
       : new EvmAddress(issuerId);
 
-    const res = await handler.addIssuer(
+    const res = await this.queryAdapter.isIssuer(
       securityEvmAddress,
       issuerEvmAddress,
-      securityId,
     );
 
-    return Promise.resolve(
-      new AddIssuerCommandResponse(res.error === undefined, res.id!),
-    );
+    return new IsIssuerQueryResponse(res);
   }
 }
